+++ tests/wpunit/bootstrap.php
@@ -48,7 +48,7 @@ if ( file_exists( __DIR__ . '/inc/fixtures.php' ) ) {
 }
 
 # Load the class autoloader
-require_once __DIR__ . '/../../class.e20r-members-list.php';
+require_once __DIR__ . '/../../class-e20r-members-list.php';
 
 // Include the class for PluginTestCase
 require_once __DIR__ . '/inc/PluginTestCase.php';